Add nxml-mode
This XML-editing mode requires the use of RELAX NG schemas instead of
DTDs.  But it's a pretty nice mode, and RELAX NG schemas are nicer than
W3C XML Schemas, so it's not a burden.

To convert a DTD to a RELAX NG schema, download trang and run:

  java -jar $TRANG_DIR/trang.jar foo.dtd foo.rnc

From there, assign the schema to an XML file using C-c C-s C-f.

Note that I'm not turning this on by default yet.